Refactor to run as an SmtpRecieveAgent.
